refact: refactored the dtos structure, including subfolders for organization
