+++ src/main/java/dev/dluks/minervamoney/dtos/user/RegisterUserRequestDTO.java
@@ -1,4 +1,4 @@
-package dev.dluks.minervamoney.dtos;
+package dev.dluks.minervamoney.dtos.user;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Pattern;
